requirements.txt
Django==2.2.1
django-pdfkit==0.3.1
djangorestframework==3.10.2
djangorestframework-simplejwt==4.3.0
pdfkit==0.6.0
PyJWT==1.7.1
pytz==2019.2
sqlparse==0.3.0
